# Conflicts:
#	CMakeLists.txt
#	CMakeLists.txt.user
#	Config/cubo.xml
#	src/Modules/BanknotePositionProvider.cpp
#	src/Modules/BlobFilter.cpp
#	src/Modules/Camera.cpp
#	src/Modules/EstimatePosition.cpp
#	src/Modules/PreviousBanknoteCheck.cpp
#	src/Modules/PreviousBanknoteCheck.h
#	src/Tools/File.cpp
#	src/Tools/SystemCall.cpp

+++ Utils/CalibratorTool/Views/ImageView.cpp
@@ -256,14 +256,14 @@ QMenu* ImageWidget::createFileMenu() const
   
   for (const auto &debugDrawing : drawingManager.drawings) {
     const std::string &name = debugDrawing.first;
-    int pos = name.find_first_of(":");
+    int pos = (int) name.find_first_of(":");
     std::string type = name.substr(0,pos);
     if (type.compare("representation") == 0) {
       representationDrawings.push_back(Drawing(name, name.substr(pos+1)));
     }
     else if(type.compare("module") == 0)
     {
-      int pos2 = name.find_first_of(":",pos+1);
+      int pos2 = (int) name.find_first_of(":",pos+1);
       if (pos2 > name.size()) {
          modulesDrawings[name.substr(pos+1)] = std::vector<Drawing>();
       }
